🎨 style(reviewer_submit_view.dart): refactor ReviewerSubmitView widget and add ReviewerSubmitBottomNavbar widget ✨ feat(reviewer_submit_botton_navbar.dart): add ReviewerSubmitBottomNavbar widget 🐛 fix(reviewer_submit_response.dart): add elevation to Card widget The ReviewerSubmitView widget is refactored to improve readability and maintainability. The ReviewerSubmitBottomNavbar widget is added to separate the bottom navigation bar from the main widget. The ReviewerSubmitResponse widget has a bug where the Card widget does not have an elevation, which is fixed by adding an elevation of 6.
