+++ lib/app/modules/reviewer_submit/views/reviewer_submit_view.dart
@@ -5,6 +5,7 @@ import 'dart:convert';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_agunan.dart';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_analys_response.dart';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_bisnis.dart';
+import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_botton_navbar.dart';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_gallery.dart';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_inputan.dart';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_karakter.dart';
@@ -29,7 +30,7 @@ import '../controllers/reviewer_submit_controller.dart';
 
 // ignore: must_be_immutable
 class ReviewerSubmitView extends GetView<ReviewerSubmitController> {
-  ReviewerSubmitView({Key? key}) : super(key: key);
+  const ReviewerSubmitView({Key? key}) : super(key: key);
 
   String formatDatetime(DateTime date) {
     return DateFormat('dd MMMM yyyy').format(date);
@@ -120,15 +121,26 @@ class ReviewerSubmitView extends GetView<ReviewerSubmitController> {
         );
   }
 
-  final _scrollController = ScrollController();
-
   @override
   Widget build(BuildContext context) {
+    ScrollController scrollController = ScrollController();
+
+    RxBool showButton = false.obs;
+
+    scrollController.addListener(() {
+      if (scrollController.position.pixels ==
+          scrollController.position.maxScrollExtent) {
+        // User reached the bottom of the page
+        showButton.value = true;
+      }
+    });
+
     return Scaffold(
       backgroundColor: secondaryColor,
       body: SafeArea(
         child: Scrollbar(
           child: SingleChildScrollView(
+            controller: scrollController,
             child: Container(
               padding: const EdgeInsets.all(16),
               child: FormBuilder(
@@ -305,108 +317,19 @@ class ReviewerSubmitView extends GetView<ReviewerSubmitController> {
                       subtitleStyle: subtitleStyle(),
                     ),
                     const SizedBox(height: 20),
-                    GFButton(
-                      onPressed: () {
-                        if (controller.formKey.currentState!
-                            .saveAndValidate()) {
-                          // controller.submit();
-                          debugPrint(controller.formKey.currentState!.value
-                              .toString());
-                          Get.dialog(
-                            AlertDialog(
-                              title: const Text(
-                                'Submit',
-                                style: TextStyle(
-                                  fontSize: 20,
-                                  fontWeight: FontWeight.bold,
-                                ),
-                              ),
-                              content: const Text(
-                                'Dengan menekan tombol Ya, data diatas akan dikirim ke pemutus yang dipilih, dan status pengajuan berubah menjadi REVIEWED. Apakah anda yakin?',
-                                style: TextStyle(
-                                  fontSize: 15,
-                                ),
-                              ),
-                              actions: [
-                                GFButton(
-                                  color: GFColors.DANGER,
-                                  size: GFSize.LARGE,
-                                  onPressed: () {
-                                    Navigator.pop(context);
-                                  },
-                                  child: const Text('Tidak'),
-                                ),
-                                GFButton(
-                                  color: GFColors.SUCCESS,
-                                  size: GFSize.LARGE,
-                                  onPressed: () {
-                                    var list =
-                                        controller.formKey.currentState!.value;
-
-                                    // Transform map to list
-                                    var list2 = list.entries.toList();
-
-                                    // // remove MapEntry and key
-                                    list2.removeWhere(
-                                      (element) =>
-                                          element.key == 'pemutus' ||
-                                          element.key == 'tglReview' ||
-                                          element.key == 'inputan' ||
-                                          element.key == 'keuangan' ||
-                                          element.key == 'karakter' ||
-                                          element.key == 'bisnis' ||
-                                          element.key == 'usaha' ||
-                                          element.key == 'agunan' ||
-                                          element.key == 'berkas',
-                                    );
-
-                                    // debugPrint('list2: $list2');
-
-                                    // Transform list2 to list of string
-                                    var list3 =
-                                        list2.map((e) => e.value).toList();
-
-                                    // list3.removeWhere((element) => element.k)
-
-                                    // transform list3 to string
-                                    list3 =
-                                        list3.map((e) => e.toString()).toList();
-
-                                    controller.bahasanReviewer = list3;
-
-                                    var listFinal = controller.bahasanReviewer;
-
-                                    debugPrint(listFinal.toString());
-
-                                    Navigator.pop(context);
-                                    controller.saveReview();
-                                  },
-                                  child: const Text('Ya'),
-                                ),
-                              ],
-                            ),
-                          );
-                        } else {
-                          debugPrint('validation failed');
-                        }
-                      },
-                      text: 'Submit',
-                      shape: GFButtonShape.square,
-                      color: GFColors.SUCCESS,
-                      fullWidthButton: true,
-                      size: GFSize.LARGE,
-                      textStyle: const TextStyle(
-                        color: Colors.white,
-                        fontSize: 18,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
                   ],
                 ),
               ),
             ),
           ),
         ),
+      ),
+      bottomNavigationBar: Obx(
+        () => showButton.value
+            ? ReviewerSubmitBottomNavbar(
+                controller: controller,
+              )
+            : const SizedBox.shrink(),
       ),
     );
   }
